+++ pa/Sample3.java
@@ -1,0 +1,10 @@
+package pa;
+
+class Sample3
+{
+  public static void main(String[] args)
+  {
+    Car car1 = new Car();
+    car1.show();
+  }
+}
